Refactor state interface
