+++ .vscode/settings.json
@@ -12,5 +12,8 @@
     "buildTarget" : {
       "visibility": "hidden"
                },
-     },
+    },
+    "files.associations": {
+      "cstdlib": "cpp"
+    },
 }
