+++ rockblock/src/pico/rock_machine_mobuffer_state.cpp
@@ -11,7 +11,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-#include "rock_machine.hpp"
+#include "rock_machine_state.hpp"
 
 #define SBDWT "AT+SBDWT="
 #define SBDWT_TIMEOUT 10000  // message timeout (ms)
