+++ rockblock/src/pico/rock_machine_rts_state.cpp
@@ -11,7 +11,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-#include "rock_machine.hpp"
+#include "rock_machine_state.hpp"
 
 #define RTS_CMD "AT\r"
 #define RTS_TIMEOUT 10000  // RTS timeout (ms)
