+++ rockblock/src/pico/rock_machine_state.cpp
@@ -10,7 +10,12 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-#include "rock_machine.hpp"
+#include "rock_machine_state.hpp"
+
+void rock_machine_state::start(rock_machine*) {}
+void rock_machine_state::send(rock_machine*) {}       
+void rock_machine_state::send_ok(rock_machine*, char* response) {}
+void rock_machine_state::repeat(rock_machine*) {}
 
 // Change the machine state
 void rock_machine_state::change_state(rock_machine* rock, rock_machine_state* state) {
